refactor(editor): Import DEFAULT_SETTINGS from constants directly
Point all consumers at the canonical @/app/constants/workflows location
and drop the back-compat re-export from useWorkflowDocumentSettings.
